feat: moved popup to invite member component
